requirements.txt
# Pinned for reproducible installs on Python 3.10+ (including 3.13/3.14 where wheels exist).
# Bump these intentionally when upgrading; run the app and a sample pipeline after changes.

streamlit==1.56.0
pandas==2.3.3
numpy==2.4.2
plotly==6.0.1
Pillow==12.2.0
python-dotenv==1.1.0
